requirements-dev.txt
ipython
ipdb
mock
responses

pylint
coverage
django_nose
pinocchio
nose-html-reporting
fake-factory
